Academic Student Succsess Machine Learning/deployment/requirements.txt
pandas== 2.0.3
numpy== 1.24.3
seaborn== 0.12.2
matplotlib== 3.7.2
scikit-learn== 1.3.0
imbalanced-learn== 0.11.0
streamlit_extras
feature_engine